# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation with Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base"
        },
        {
            "url": "/api",
            "description": "API Server RRHH"
        }
    ],
    "paths": {
...
        "/docentes-contratados": {
            "get": {
                "tags": [
                    "DocentesContratados"
                ],
                "summary": "Lista paginada de docentes contratados",
                "operationId": "da1ea698ca1c82113c89ac13dd913ffd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 25
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Búsqueda sobre DNI, nombre o resolución",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DocenteContratado"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 83
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "DocentesContratados"
                ],
                "summary": "Crear docente contratado",
                "operationId": "storeDocenteContratado",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocenteContratado"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Docente creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocenteContratado"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/docentes-contratados/{id}": {
            "get": {
                "tags": [
                    "DocentesContratados"
                ],
                "summary": "Detalle de un docente contratado con sus contratos agrupados por ciclo",
                "operationId": "showDocenteContratado",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del docente contratado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Docente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "dni": {
                                                    "type": "string"
                                                },
                                                "nombres": {
                                                    "type": "string"
                                                },
                                                "contratos_por_ciclo": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "ciclo": {
                                                                "properties": {
                                                                    "anio": {
                                                                        "type": "integer"
                                                                    },
                                                                    "semestre": {
                                                                        "type": "string"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "contratos": {
                                                                "type": "array",
                                                                "items": {
                                                                    "$ref": "#/components/schemas/ContratoDocente"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "DocentesContratados"
                ],
                "summary": "Actualizar docente contratado",
                "operationId": "updateDocenteContratado",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocenteContratado"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Docente actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocenteContratado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "DocentesContratados"
                ],
                "summary": "Eliminar docente contratado",
                "operationId": "deleteDocenteContratado",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/docentes-contratados/import": {
            "post": {
                "tags": [
                    "DocentesContratados"
                ],
                "summary": "Importar Excel de contratos",
                "operationId": "importContratosDocentes",
                "parameters": [
                    {
                        "name": "anio",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    },
                    {
                        "name": "semestre",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "I",
                                "II"
                            ],
                            "example": "I"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Archivo XLSX/XLS/CSV",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Importación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Archivo importado correctamente para 2025-I"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ContratoDocente": {
                "required": [
                    "docente_id",
                    "ciclo_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "docente_id": {
                        "type": "integer"
                    },
                    "ciclo_id": {
                        "type": "integer"
                    },
                    "categoria": {
                        "type": "string",
                        "nullable": true
                    },
                    "regimen_dedicacion": {
                        "type": "string",
                        "nullable": true
                    },
                    "fecha_ingreso": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "fecha_fin": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "resolucion": {
                        "type": "string",
                        "nullable": true
                    },
                    "facultad_id": {
                        "type": "integer"
                    },
                    "programa_id": {
                        "type": "integer"
                    },
                    "tipo_id": {
                        "description": "FK a tipo_contrato_docente",
                        "type": "integer",
                        "nullable": true
                    },
                    "comentario": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "DocenteContratado": {
                "required": [
                    "dni",
                    "apellido_paterno",
                    "apellido_materno",
                    "nombres"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 101
                    },
                    "dni": {
                        "type": "string",
                        "example": "42900502"
                    },
                    "apellido_paterno": {
                        "type": "string",
                        "example": "ALVA"
                    },
                    "apellido_materno": {
                        "type": "string",
                        "example": "ZAPATA"
                    },
                    "nombres": {
                        "type": "string",
                        "example": "JULIANA DEL PILAR"
                    },
                    "sexo": {
                        "type": "string",
                        "enum": [
                            "M",
                            "F",
                            "X"
                        ],
                        "nullable": true
                    },
                    "mayor_grado_academico": {
                        "type": "string",
                        "example": "Maestro",
                        "nullable": true
                    },
                    "correo": {
                        "type": "string",
                        "example": "juliana.alva@email.com",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "987654321",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Pega tu token así: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}